+++ src/CommonMethods.ts
